refactor: initial separation to working nogood
