+++ pumpkin-crates/conflict-resolvers/src/resolvers/mod.rs
@@ -1,5 +1,7 @@
 //! Contains the conflict resolvers.
 mod no_learning_resolver;
 mod resolution_resolver;
+mod working_nogood;
 pub use no_learning_resolver::*;
 pub use resolution_resolver::*;
+pub(crate) use working_nogood::*;
