+++ pumpkin-crates/conflict-resolvers/src/resolvers/resolution_resolver.rs
@@ -1,31 +1,23 @@
 use pumpkin_core::asserts::pumpkin_assert_advanced;
-use pumpkin_core::asserts::pumpkin_assert_moderate;
-use pumpkin_core::asserts::pumpkin_assert_simple;
 use pumpkin_core::conflict_resolving::AnalysisMode;
 use pumpkin_core::conflict_resolving::ConflictAnalysisContext;
 use pumpkin_core::conflict_resolving::ConflictResolver;
-use pumpkin_core::containers::HashMap;
-use pumpkin_core::containers::KeyValueHeap;
-use pumpkin_core::containers::StorageKey;
 use pumpkin_core::create_statistics_struct;
 use pumpkin_core::predicates::Lbd;
 use pumpkin_core::predicates::Predicate;
 use pumpkin_core::predicates::PredicateIdGenerator;
-use pumpkin_core::propagation::PredicateId;
 use pumpkin_core::propagation::ReadDomains;
 use pumpkin_core::state::CurrentNogood;
 use pumpkin_core::statistics::Statistic;
 use pumpkin_core::statistics::StatisticLogger;
 use pumpkin_core::statistics::moving_averages::CumulativeMovingAverage;
 use pumpkin_core::statistics::moving_averages::MovingAverage;
-use pumpkin_core::variables::DomainId;
 
-use crate::minimisers::IterativeMinimiser;
 use crate::minimisers::NogoodMinimiser;
-use crate::minimisers::ProcessingResult;
 use crate::minimisers::RecursiveMinimiser;
 use crate::minimisers::SemanticMinimisationMode;
 use crate::minimisers::SemanticMinimiser;
+use crate::resolvers::WorkingNogood;
 
 /// [`ConflictResolver`] which resolves conflicts according to the CDCL procedure.
 ///
@@ -42,31 +34,16 @@ use crate::minimisers::SemanticMinimiser;
 /// Handbook of satisfiability, pp. 131–153, 2009.
 #[derive(Clone, Debug)]
 pub struct ResolutionResolver {
-    /// Heap containing the predicates which still need to be processed; sorted non-increasing
-    /// based on trail-index where implied predicates are processed first.
-    to_process_heap: KeyValueHeap<PredicateId, u32>,
+    working_nogood: WorkingNogood,
     /// The generator is used in combination with the heap to keep track of which predicates are
     /// stored in the heap.
     predicate_id_generator: PredicateIdGenerator,
-    /// Predicates which have been processed and have been determined to be (potentially) part of
-    /// the nogood.
-    ///
-    /// Note that this structure may contain duplicates which are removed at the end by semantic
-    /// minimisation.
-    processed_nogood_predicates: Vec<Predicate>,
     /// The type of learning that the resolver employs (e.g., 1UIP, All-decision).
     mode: AnalysisMode,
     /// Re-usable buffer which reasons are written into.
     reason_buffer: Vec<Predicate>,
     /// Computes the LBD for nogoods.
     lbd_helper: Lbd,
-    /// A helper for keeping track of how many [`Predicate`]s concerning a specific [`DomainId`]
-    /// are present in the working nogood.
-    ///
-    /// This is used when determining when to stop resolving when using CPIP learning (see
-    /// [`AnalysisMode::CPIP`]).
-    unique_variable_helper: HashMap<DomainId, u32>,
-
     /// A minimiser which recursively determines whether a predicate is redundant in the nogood.
     recursive_minimiser: RecursiveMinimiser,
     /// A minimiser which determines whether a predicate is redundant in the nogood based on its
@@ -80,13 +57,6 @@ pub struct ResolutionResolver {
     ///
     /// Note that semantic minimisation is always applied to remove duplicates.
     should_minimise: bool,
-
-    /// Whether to perform iterative minimisation.
-    ///
-    /// Iterative minimisation is semantic minimisation applied *while* resolving.
-    iterative_minimisation: bool,
-    /// The structure used for iterative minimisation.
-    iterative_minimiser: IterativeMinimiser,
 }
 
 impl Default for ResolutionResolver {
@@ -99,7 +69,6 @@ create_statistics_struct!(
     /// The statistics related to clause learning
     LearnedNogoodStatistics {
         nogood_statistics: NogoodStatistics,
-        iterative_minimisation_statistics: IterativeMinimisationStatistics,
         cpip_statistics: CpipStatistics,
 });
 
@@ -126,41 +95,13 @@ create_statistics_struct!(CpipStatistics {
         num_regular_nogood_learned: usize,
 });
 
-create_statistics_struct!(IterativeMinimisationStatistics {
-    /// The number of removed predicates by iterative minimisation.
-    num_removed: usize,
-    /// The number of removed predicates from the current decision level.
-    num_removed_current_decision_level: usize,
-    /// The number of removed predicates from the previous decision level.
-    num_removed_previous_decision_level: usize
-});
-
-/// Indicates whether a [`Predicate`] is redundant or non-redundant when adding it to the nogood.
-#[derive(Clone, Copy, Debug)]
-enum IterativeRedundancyStatus {
-    Redundant,
-    NonRedundant,
-}
-
-/// Indicates whether a [`Predicate`] was replaced during iterative minimisation.
-///
-/// If we have the case that the predicates [x >= v] and [x <= v] are present in the nogood, then
-/// we can potentially replace it with [x == v]. However, if [x == v] would be the next predicate
-/// to be resolved upon, then we cannot add it to the nogood. Hence, we indicate whether this
-/// replacement is possible using this enum.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ReplacementStatus {
-    NotReplaced,
-    Replaced,
-}
-
 impl ConflictResolver for ResolutionResolver {
     fn resolve_conflict(&mut self, context: &mut ConflictAnalysisContext) {
         self.learn_nogood(context);
 
         let lbd = self
             .lbd_helper
-            .compute_lbd(&self.processed_nogood_predicates, context);
+            .compute_lbd(&self.working_nogood.processed_nogood_predicates, context);
 
         // Update statistics
         self.statistics
@@ -168,14 +109,14 @@ impl ConflictResolver for ResolutionResolver {
             .average_lbd
             .add_term(lbd as u64);
         self.statistics.nogood_statistics.num_unit_nogoods_learned +=
-            (self.processed_nogood_predicates.len() == 1) as u64;
+            (self.working_nogood.processed_nogood_predicates.len() == 1) as u64;
         self.statistics
             .nogood_statistics
             .average_learned_nogood_length
-            .add_term(self.processed_nogood_predicates.len() as u64);
+            .add_term(self.working_nogood.processed_nogood_predicates.len() as u64);
 
         let backtrack_level = context.process_learned_nogood(
-            self.processed_nogood_predicates.clone(),
+            self.working_nogood.processed_nogood_predicates.clone(),
             lbd,
             self.mode.uses_cpip(),
         );
@@ -190,13 +131,7 @@ impl ConflictResolver for ResolutionResolver {
         self.statistics
             .nogood_statistics
             .log(statistic_logger.clone());
-        if self.iterative_minimisation {
-            self.statistics
-                .iterative_minimisation_statistics
-                .log(statistic_logger.clone());
-            self.iterative_minimiser
-                .log_statistics(statistic_logger.clone());
-        }
+        self.working_nogood.log_statistics(statistic_logger.clone());
         if self.mode.uses_cpip() {
             self.statistics
                 .cpip_statistics
@@ -213,18 +148,21 @@ impl ResolutionResolver {
     pub fn new(mode: AnalysisMode, should_minimise: bool, iterative_minimisation: bool) -> Self {
         Self {
             mode,
-            to_process_heap: Default::default(),
             predicate_id_generator: Default::default(),
-            processed_nogood_predicates: Default::default(),
             reason_buffer: Default::default(),
             lbd_helper: Default::default(),
             recursive_minimiser: Default::default(),
             semantic_minimiser: Default::default(),
             statistics: Default::default(),
             should_minimise,
-            iterative_minimisation,
-            iterative_minimiser: Default::default(),
-            unique_variable_helper: Default::default(),
+            working_nogood: WorkingNogood {
+                to_process_heap: Default::default(),
+                processed_nogood_predicates: Default::default(),
+                unique_variable_helper: Default::default(),
+                iterative_minimisation,
+                iterative_minimiser: Default::default(),
+                iterative_minimisation_statistics: Default::default(),
+            },
         }
     }
 
@@ -235,12 +173,20 @@ impl ResolutionResolver {
 
         // Initialise the data structures with the conflict nogood.
         for predicate in conflict_nogood.iter() {
-            self.add_predicate_to_conflict_nogood(*predicate, self.mode, context);
+            self.working_nogood.add_predicate_to_conflict_nogood(
+                *predicate,
+                self.mode,
+                context,
+                &mut self.predicate_id_generator,
+            );
         }
 
         // Record conflict nogood size statistics.
-        let num_initial_conflict_predicates =
-            self.to_process_heap.num_nonremoved_elements() + self.processed_nogood_predicates.len();
+        let num_initial_conflict_predicates = self
+            .working_nogood
+            .to_process_heap
+            .num_nonremoved_elements()
+            + self.working_nogood.processed_nogood_predicates.len();
         self.statistics
             .nogood_statistics
             .average_conflict_size
@@ -258,17 +204,21 @@ impl ResolutionResolver {
         // <= v]). In this case there will be two predicates left from the current decision
         // level, and both will be decisions. This is accounted for below.
         while self.mode.should_continue_resolving(
-            &self.to_process_heap,
+            &self.working_nogood.to_process_heap,
             &mut self.predicate_id_generator,
-            &mut self.unique_variable_helper,
+            &mut self.working_nogood.unique_variable_helper,
         ) {
             // Replace the predicate from the nogood that has been assigned last on the trail.
             //
             // This is done in two steps:
             // 1) Pop the predicate last assigned on the trail from the nogood.
-            let next_predicate = self.pop_predicate_from_conflict_nogood();
-            if self.iterative_minimisation {
-                self.iterative_minimiser.remove_predicate(next_predicate);
+            let next_predicate = self
+                .working_nogood
+                .pop_predicate_from_conflict_nogood(&mut self.predicate_id_generator, self.mode);
+            if self.working_nogood.iterative_minimisation {
+                self.working_nogood
+                    .iterative_minimiser
+                    .remove_predicate(next_predicate);
             }
 
             // 2) Get the reason for the predicate and add it to the nogood.
@@ -277,15 +227,20 @@ impl ResolutionResolver {
             let _ = context.get_propagation_reason(
                 next_predicate,
                 CurrentNogood::new(
-                    &self.to_process_heap,
-                    &self.processed_nogood_predicates,
+                    &self.working_nogood.to_process_heap,
+                    &self.working_nogood.processed_nogood_predicates,
                     &self.predicate_id_generator,
                 ),
                 &mut self.reason_buffer,
             );
 
             for i in 0..self.reason_buffer.len() {
-                self.add_predicate_to_conflict_nogood(self.reason_buffer[i], self.mode, context);
+                self.working_nogood.add_predicate_to_conflict_nogood(
+                    self.reason_buffer[i],
+                    self.mode,
+                    context,
+                    &mut self.predicate_id_generator,
+                );
             }
         }
 
@@ -294,397 +249,8 @@ impl ResolutionResolver {
 
     /// Clears all data structures to prepare for the new conflict analysis.
     fn clean_up(&mut self) {
-        self.processed_nogood_predicates.clear();
         self.predicate_id_generator.clear();
-        self.to_process_heap.clear();
-
-        // TODO: make more efficient
-        if self.iterative_minimisation {
-            self.iterative_minimiser.clear();
-        }
-        self.unique_variable_helper.clear();
-    }
-
-    /// Add the predicate to the current conflict nogood if we know it needs to be added.
-    ///
-    /// If a `root_explanation_context` is provided, then root-level assignments are explained as
-    /// well in the proof log.
-    fn add_predicate_to_conflict_nogood(
-        &mut self,
-        predicate: Predicate,
-        mode: AnalysisMode,
-        context: &mut ConflictAnalysisContext,
-    ) {
-        let dec_level = context
-            .get_checkpoint_for_predicate(predicate)
-            .unwrap_or_else(|| {
-                panic!(
-                    "Expected predicate {predicate} to be assigned but bounds were ({}, {})",
-                    context.lower_bound(&predicate.get_domain()),
-                    context.lower_bound(&predicate.get_domain()),
-                )
-            });
-        // Ignore root level predicates.
-        if dec_level == 0 {
-            context.explain_root_assignment(predicate);
-
-            if self.iterative_minimisation {
-                self.iterative_minimiser.apply_predicate(predicate, context);
-            }
-        }
-        // 1UIP
-        // If the variables are from the current decision level then we want to potentially add
-        // them to the heap, otherwise we add it to the predicates from lower-decision levels
-        //
-        // All-decision Learning
-        // If the variables are not decisions then we want to potentially add them to the heap,
-        // otherwise we add it to the decision predicates which have been discovered previously
-        else if mode.predicate_should_be_processed(predicate, dec_level, context) {
-            let predicate_id = self.predicate_id_generator.get_id(predicate);
-            // The first time we encounter the predicate, we initialise its value in the
-            // heap.
-            //
-            // Note that if the predicate is already in the heap, no action needs to be
-            // taken. It can happen that a predicate is returned
-            // multiple times as a reason for other predicates.
-
-            // TODO: could improve the heap structure to be more user-friendly.
-
-            // Here we manually adjust the size of the heap to accommodate new elements.
-            while self.to_process_heap.len() <= predicate_id.index() {
-                let next_id = PredicateId::create_from_index(self.to_process_heap.len());
-                self.to_process_heap.grow(next_id, 0);
-                self.to_process_heap.delete_key(next_id);
-            }
-
-            // Then we check whether the predicate was not already present in the heap, if
-            // this is not the case then we insert it
-            if !self.to_process_heap.is_key_present(predicate_id)
-                && *self.to_process_heap.get_value(predicate_id) == 0
-            {
-                if let IterativeRedundancyStatus::Redundant =
-                    self.check_for_iterative_redundancy(predicate, context, predicate_id)
-                {
-                    if dec_level == context.get_checkpoint() {
-                        self.statistics
-                            .iterative_minimisation_statistics
-                            .num_removed_current_decision_level += 1
-                    } else {
-                        self.statistics
-                            .iterative_minimisation_statistics
-                            .num_removed_previous_decision_level += 1
-                    }
-                    return;
-                }
-
-                context.predicate_appeared_in_conflict(predicate);
-
-                // The goal is to traverse predicate in reverse order of the trail.
-                //
-                // However some predicates may share the trail position. For example, if a
-                // predicate that was posted to trail resulted in
-                // some other predicates being true, then all
-                // these predicates would have the same trail position.
-                //
-                // When considering the predicates in reverse order of the trail, the
-                // implicitly set predicates are posted after the
-                // explicitly set one, but they all have the same
-                // trail position.
-                //
-                // To remedy this, we make a tie-breaking scheme to prioritise implied
-                // predicates over explicit predicates. This is done
-                // by assigning explicitly set predicates the
-                // value `2 * trail_position`, whereas implied predicates get `2 *
-                // trail_position + 1`.
-                let heap_value = get_heap_value(predicate, context);
-
-                // We restore the key and since we know that the value is 0, we can safely
-                // increment with `heap_value`
-                self.to_process_heap.restore_key(predicate_id);
-                self.to_process_heap.increment(predicate_id, heap_value);
-                mode.add_predicate_to_nogood(predicate, &mut self.unique_variable_helper);
-
-                pumpkin_assert_moderate!(
-                    *self.to_process_heap.get_value(predicate_id) == heap_value,
-                    "The value in the heap should be the same as was added"
-                )
-            }
-        } else {
-            // We do not check for duplicate, we simply add the predicate.
-            // Semantic minimisation will later remove duplicates and do other processing.
-            self.processed_nogood_predicates.push(predicate);
-        }
-    }
-
-    /// Checks whether the provided [`Predicate`] is redundant given the current working nogood.
-    fn check_for_iterative_redundancy(
-        &mut self,
-        predicate: Predicate,
-        context: &mut ConflictAnalysisContext<'_>,
-        predicate_id: PredicateId,
-    ) -> IterativeRedundancyStatus {
-        if !self.iterative_minimisation {
-            return IterativeRedundancyStatus::NonRedundant;
-        }
-
-        // We ask the iterative minimiser the status of the predicate.
-        let process_predicate = self
-            .iterative_minimiser
-            .process_predicate(predicate, context);
-
-        // Based on the status, we proceed accordingly.
-        match process_predicate {
-            ProcessingResult::Redundant => {
-                // The provided predicate is redundant.
-                //
-                // The key is not currently present, but it has been assigned a value; we need
-                // to reset that value to 0.
-                if predicate_id.index() < self.to_process_heap.len() {
-                    self.to_process_heap.set_value(predicate_id, 0);
-                }
-                // Then we delete the key if it was present.
-                self.to_process_heap.delete_key(predicate_id);
-
-                self.statistics
-                    .iterative_minimisation_statistics
-                    .num_removed += 1;
-
-                // We know that the element is redundant, so we can indicate that it does not need
-                // to be processed.
-                IterativeRedundancyStatus::Redundant
-            }
-            ProcessingResult::ReplacedPresent { removed } => {
-                // First, we remove the predicates.
-                self.remove_predicates(removed);
-
-                // Then we apply the provided predicate to the domain after removing all of the
-                // previous predicates.
-                self.iterative_minimiser.apply_predicate(predicate, context);
-
-                // We also know that the provided predicate is not redundant so we can add it to
-                // the nogood.
-                IterativeRedundancyStatus::NonRedundant
-            }
-            ProcessingResult::PossiblyReplacedWithNew {
-                potentially_removed: previous,
-                new_predicate,
-                removed,
-            } => {
-                // Adding the new predicate would lead it to be replaced with another predicate
-                // (e.g. we are adding [x >= 5] and run into the situation that [x >= 5] /\ [x != 5]
-                // -> [x >= 6] occurs; in this case, the removed predicate would be [x != 5] and the
-                // new_predicate would be [x >= 6]).
-                //
-                // It is important to not enter a loop here due to implied predicates. In the
-                // previous example, if the next predicate to be resolved upon is [x >= 6], then a
-                // loop could be entered in which [x >= 5] /\ [x != 5] are merged into [x >= 6],
-                // after which the reason for [x >= 6] is [x >= 5] /\ [x != 5] -> [x >= 6], which
-                // are then merged back, etc.
-                //
-                // To resolve this issue, we only replace the provided predicate when the new
-                // predicate would not be the next one to be resolved upon.
-                self.statistics
-                    .iterative_minimisation_statistics
-                    .num_removed += 1;
-
-                // First, we remove the predicates that are removed either way.
-                self.remove_predicates(removed);
-
-                // We split into two cases:
-                // 1. The new predicate is of the current decision level.
-                // 2. The new predicate is of a previous decision level.
-                if context.get_checkpoint_for_predicate(new_predicate).unwrap()
-                    == context.get_checkpoint()
-                {
-                    // Next, we check whether we can replace the elements with `new_predicate`.
-                    if ReplacementStatus::Replaced
-                        == self.replace_if_possible_current_level(context, previous, new_predicate)
-                    {
-                        self.to_process_heap.set_value(predicate_id, 0);
-                        self.to_process_heap.delete_key(predicate_id);
-
-                        // We can replace the elements with `new_predicate`, so we indicate that we
-                        // do not need to add `predicate`.
-                        IterativeRedundancyStatus::Redundant
-                    } else {
-                        // We cannot replace the elements, so we add `predicate` to the iterative
-                        // minimiser.
-                        self.iterative_minimiser.apply_predicate(predicate, context);
-
-                        // And we indicate that we need to add `predicate` to the nogood.
-                        IterativeRedundancyStatus::NonRedundant
-                    }
-                } else {
-                    // If `new_predicate` is from a previous decision level, then we can always
-                    // replace the elements, so we do this directly.
-                    self.replace_previous_level(context, previous, new_predicate);
-
-                    self.to_process_heap.set_value(predicate_id, 0);
-                    self.to_process_heap.delete_key(predicate_id);
-
-                    // And we indicate that we do not need to add `predicate` to the nogood.
-                    IterativeRedundancyStatus::Redundant
-                }
-            }
-            ProcessingResult::NotRedundant => {
-                // `predicate` is not redundant and we can add it directly to the nogood.
-                self.iterative_minimiser.apply_predicate(predicate, context);
-                IterativeRedundancyStatus::NonRedundant
-            }
-        }
-    }
-
-    /// Removes the provided predicates from the predicates to be resolved upon, or the ones
-    /// already in the nogood.
-    fn remove_predicates(&mut self, removed: Vec<Predicate>) {
-        // The provided predicate has replaced a multitude of other predicates -> we need
-        // to remove all of these predicates.
-        //
-        // Hence, we go over all of the removed predicates.
-        for removed_predicate in removed {
-            self.statistics
-                .iterative_minimisation_statistics
-                .num_removed += 1;
-
-            // And we also remove it from the iterative minimiser itself.
-            self.iterative_minimiser.remove_predicate(removed_predicate);
-
-            let removed_id = self.predicate_id_generator.get_id(removed_predicate);
-            // We differentiate between two cases:
-            // 1. The removed predicate is from the current decision level and we need to remove it
-            //    from the heap.
-            // 2. The removed predicate is from the previous decision level, and we remove it from
-            //    there.
-            if self.to_process_heap.is_key_present(removed_id) {
-                self.mode.remove_predicate_from_nogood(
-                    removed_predicate,
-                    &mut self.unique_variable_helper,
-                );
-                // The key is not currently present, but it has been assigned a value; we
-                // need to reset that value to 0.
-                if removed_id.index() < self.to_process_heap.len() {
-                    self.to_process_heap.set_value(removed_id, 0);
-                }
-                self.to_process_heap.delete_key(removed_id);
-            } else if let Some(position) = self
-                .processed_nogood_predicates
-                .iter()
-                .position(|predicate| *predicate == removed_predicate)
-            {
-                let _ = self.processed_nogood_predicates.remove(position);
-            }
-        }
-    }
-
-    /// Replaces the provided `element` with `new_predicate` if `new_predicate` would not be the
-    /// next element to be resolved upon.
-    fn replace_if_possible_current_level(
-        &mut self,
-        context: &mut ConflictAnalysisContext<'_>,
-        element: Predicate,
-        new_predicate: Predicate,
-    ) -> ReplacementStatus {
-        // We first calculate the value in the heap of the new_predicate.
-        let heap_value = get_heap_value(new_predicate, context);
-
-        // Then we check whether this is a lower value than the current maximum in the heap.
-        //
-        // If it has a higher value than the current maximum in the heap, then we check whether
-        // there is only a single element in the heap, and it is the `element` that we are trying
-        // to replace (i.e., replacing `element` with `new_predicate` would make `new_predicate`
-        // the asserting atomic constraint).
-        if heap_value
-            < self
-                .to_process_heap
-                .peek_max()
-                .map(|(_, value)| *value)
-                .unwrap_or_default()
-            || (self.to_process_heap.num_nonremoved_elements() == 1
-                && self
-                    .to_process_heap
-                    .peek_max()
-                    .map(|(&predicate_id, _)| {
-                        predicate_id == self.predicate_id_generator.get_id(element)
-                    })
-                    .unwrap())
-        {
-            // If it is, then we can safely replace `element` with `new_predicate`
-            //
-            // First, we remove `element`.
-            if context.get_checkpoint_for_predicate(element).unwrap() == context.get_checkpoint() {
-                let element_id = self.predicate_id_generator.get_id(element);
-                if self.to_process_heap.is_key_present(element_id) {
-                    self.mode
-                        .remove_predicate_from_nogood(element, &mut self.unique_variable_helper);
-                }
-
-                // The key is not currently present, but it has been assigned a value; we
-                // need to reset that value to 0.
-                if element_id.index() < self.to_process_heap.len() {
-                    self.to_process_heap.set_value(element_id, 0);
-                }
-                self.to_process_heap.delete_key(element_id);
-            } else {
-                if let Some(index) = self
-                    .processed_nogood_predicates
-                    .iter()
-                    .position(|predicate| *predicate == element)
-                {
-                    let _ = self.processed_nogood_predicates.remove(index);
-                }
-            }
-            self.iterative_minimiser.remove_predicate(element);
-
-            // Then we add it to the current nogood.
-            self.add_predicate_to_conflict_nogood(new_predicate, self.mode, context);
-
-            // And we return that `element` was removed.
-            return ReplacementStatus::Replaced;
-        }
-
-        // `new_predicate` would be the element to be removed, so we cannot replace `element`.
-        ReplacementStatus::NotReplaced
-    }
-
-    /// Replaces `element` with `new_predicate`, where we know that `new_predicate` and `element`
-    /// are from the previous checkpoint.
-    fn replace_previous_level(
-        &mut self,
-        context: &mut ConflictAnalysisContext<'_>,
-        element: Predicate,
-        new_predicate: Predicate,
-    ) {
-        pumpkin_assert_simple!(
-            get_heap_value(new_predicate, context)
-                < self
-                    .to_process_heap
-                    .peek_max()
-                    .map(|(_, value)| *value)
-                    .unwrap_or(u32::MAX)
-        );
-
-        // We remove `element` from the previous decision level.
-        if let Some(index) = self
-            .processed_nogood_predicates
-            .iter()
-            .position(|predicate| *predicate == element)
-        {
-            let _ = self.processed_nogood_predicates.remove(index);
-        }
-
-        // And we remove `element` from the iterative minimiser.
-        self.iterative_minimiser.remove_predicate(element);
-
-        // Next, we add the `new_predicate` to the nogood.
-        self.add_predicate_to_conflict_nogood(new_predicate, self.mode, context);
-    }
-
-    fn pop_predicate_from_conflict_nogood(&mut self) -> Predicate {
-        let next_predicate_id = self.to_process_heap.pop_max().unwrap();
-        let predicate = self.predicate_id_generator.get_predicate(next_predicate_id);
-        self.mode
-            .remove_predicate_from_nogood(predicate, &mut self.unique_variable_helper);
-        predicate
+        self.working_nogood.clean_up();
     }
 
     fn extract_final_nogood(&mut self, context: &mut ConflictAnalysisContext) {
@@ -694,9 +260,9 @@ impl ResolutionResolver {
         // Depending on what mode we are in, we first remove the elements which are remaining in
         // the heap.
         let num_removed = self.mode.remove_final_predicates(
-            &mut self.to_process_heap,
+            &mut self.working_nogood.to_process_heap,
             &mut self.predicate_id_generator,
-            &mut self.processed_nogood_predicates,
+            &mut self.working_nogood.processed_nogood_predicates,
         );
 
         self.statistics
@@ -720,61 +286,46 @@ impl ResolutionResolver {
         } else {
             SemanticMinimisationMode::DisableEqualityMerging
         });
-        self.semantic_minimiser
-            .minimise(context, &mut self.processed_nogood_predicates);
+        self.semantic_minimiser.minimise(
+            context,
+            &mut self.working_nogood.processed_nogood_predicates,
+        );
 
         if self.should_minimise {
             // Then we perform recursive minimisation to remove the dominated predicates
-            self.recursive_minimiser
-                .minimise(context, &mut self.processed_nogood_predicates);
+            self.recursive_minimiser.minimise(
+                context,
+                &mut self.working_nogood.processed_nogood_predicates,
+            );
 
             // We perform a final semantic minimisation call which allows the merging of the
             // equality predicates which remain in the nogood
             self.semantic_minimiser
                 .set_mode(SemanticMinimisationMode::EnableEqualityMerging);
-            self.semantic_minimiser
-                .minimise(context, &mut self.processed_nogood_predicates);
+            self.semantic_minimiser.minimise(
+                context,
+                &mut self.working_nogood.processed_nogood_predicates,
+            );
         }
 
         pumpkin_assert_advanced!(
-            self.processed_nogood_predicates
+            self.working_nogood
+                .processed_nogood_predicates
                 .iter()
                 .filter(|p| context.evaluate_predicate(**p) == Some(true))
                 .count()
-                >= self.processed_nogood_predicates.len() - 1,
+                >= self.working_nogood.processed_nogood_predicates.len() - 1,
             "Not all predicates evaluated to true: {:?}",
-            self.processed_nogood_predicates
+            self.working_nogood
+                .processed_nogood_predicates
                 .iter()
                 .filter(|p| context.evaluate_predicate(**p) != Some(true))
                 .collect::<Vec<_>>()
         );
 
         // TODO: asserting predicate may be bumped twice, probably not a problem.
-        for predicate in self.processed_nogood_predicates.iter() {
+        for predicate in self.working_nogood.processed_nogood_predicates.iter() {
             context.predicate_appeared_in_conflict(*predicate);
         }
-    }
-}
-
-/// Returns the value assigned to the provided [`Predicate`] in the heap.
-///
-/// The idea is that implied predicates (i.e., predicates which are not explicitly on the trail)
-/// should be resolved upon before the predicates which are explicitly on the trail.
-///
-/// Panics if the provided [`Predicate`] is not currently true on the trail.
-fn get_heap_value(predicate: Predicate, context: &mut ConflictAnalysisContext<'_>) -> u32 {
-    if context.get_state().is_on_trail(predicate) {
-        context
-            .get_state()
-            .trail_position(predicate)
-            .expect("Predicate should be true during conflict analysis") as u32
-            * 2
-    } else {
-        context
-            .get_state()
-            .trail_position(predicate)
-            .expect("Predicate should be true during conflict analysis") as u32
-            * 2
-            + 1
     }
 }
